+++ docs/arewefastyet.md
@@ -12,14 +12,15 @@ Vitess has to ensure it's delivering flawless performance to its users. In order
       --config string    config file (default is $HOME/.config/arewefastyet/config.yaml)
   -h, --help             help for arewefastyet
       --secrets string   secrets file
-  -t, --toggle           Help message for toggle
 ```
 
 ### SEE ALSO
 
+* [arewefastyet api](arewefastyet_api.md)	 - Starts the api server of arewefastyet and the CRON service
+* [arewefastyet completion](arewefastyet_completion.md)	 - Generate the autocompletion script for the specified shell
 * [arewefastyet exec](arewefastyet_exec.md)	 - Execute a task
 * [arewefastyet gen](arewefastyet_gen.md)	 - Generate things
+* [arewefastyet ghapp](arewefastyet_ghapp.md)	 - Starts the arewefastyet GitHub bot
 * [arewefastyet macrobench](arewefastyet_macrobench.md)	 - Top level command to manage macrobenchmarks
 * [arewefastyet microbench](arewefastyet_microbench.md)	 - Top level command to manage microbenchmarks
-* [arewefastyet web](arewefastyet_web.md)	 - Starts the HTTP web server
