+++ docs/arewefastyet_completion.md
@@ -1,0 +1,31 @@
+## arewefastyet completion
+
+Generate the autocompletion script for the specified shell
+
+### Synopsis
+
+Generate the autocompletion script for arewefastyet for the specified shell.
+See each sub-command's help for details on how to use the generated script.
+
+
+### Options
+
+```
+  -h, --help   help for completion
+```
+
+### Options inherited from parent commands
+
+```
+      --config string    config file (default is $HOME/.config/arewefastyet/config.yaml)
+      --secrets string   secrets file
+```
+
+### SEE ALSO
+
+* [arewefastyet](arewefastyet.md)	 - Nightly Benchmarks Project
+* [arewefastyet completion bash](arewefastyet_completion_bash.md)	 - Generate the autocompletion script for bash
+* [arewefastyet completion fish](arewefastyet_completion_fish.md)	 - Generate the autocompletion script for fish
+* [arewefastyet completion powershell](arewefastyet_completion_powershell.md)	 - Generate the autocompletion script for powershell
+* [arewefastyet completion zsh](arewefastyet_completion_zsh.md)	 - Generate the autocompletion script for zsh
+
