+++ docs/arewefastyet_completion_powershell.md
@@ -1,0 +1,38 @@
+## arewefastyet completion powershell
+
+Generate the autocompletion script for powershell
+
+### Synopsis
+
+Generate the autocompletion script for powershell.
+
+To load completions in your current shell session:
+
+	arewefastyet completion powershell | Out-String | Invoke-Expression
+
+To load completions for every new session, add the output of the above command
+to your powershell profile.
+
+
+```
+arewefastyet completion powershell [flags]
+```
+
+### Options
+
+```
+  -h, --help              help for powershell
+      --no-descriptions   disable completion descriptions
+```
+
+### Options inherited from parent commands
+
+```
+      --config string    config file (default is $HOME/.config/arewefastyet/config.yaml)
+      --secrets string   secrets file
+```
+
+### SEE ALSO
+
+* [arewefastyet completion](arewefastyet_completion.md)	 - Generate the autocompletion script for the specified shell
+
