+++ docs/arewefastyet_ghapp.md
@@ -1,0 +1,30 @@
+## arewefastyet ghapp
+
+Starts the arewefastyet GitHub bot
+
+```
+arewefastyet ghapp [flags]
+```
+
+### Options
+
+```
+      --gh-app-id int              ID of the GitHub App
+      --gh-installation-id int     GitHub installation ID of this app
+      --gh-port string             Port on which to run the github app (default "8181")
+      --gh-secret-key string       Secret key used to authenticate
+      --gh-webhook-secret string   Secrets used to verify the webhooks
+  -h, --help                       help for ghapp
+```
+
+### Options inherited from parent commands
+
+```
+      --config string    config file (default is $HOME/.config/arewefastyet/config.yaml)
+      --secrets string   secrets file
+```
+
+### SEE ALSO
+
+* [arewefastyet](arewefastyet.md)	 - Nightly Benchmarks Project
+
