+++ go/cmd/api/api.go
@@ -16,27 +16,19 @@
  * /
  */
 
-package web
+package api
 
 import (
 	"github.com/spf13/cobra"
 	"github.com/vitessio/arewefastyet/go/server"
 )
 
-func WebCmd() *cobra.Command {
+func ApiCmd() *cobra.Command {
 	var srv server.Server
 
 	cmd := &cobra.Command{
-		Use:   "web",
-		Short: "Starts the HTTP web server",
-		Long: `web command starts the HTTP web server, the credentials for which are provided as command line arguments or the configuration file. 
-It uses a MySQL and InfluxDB instances to read the metrics that will be displayed. It has an interface for macrobenchmark and microbenchmark runs.`,
-		Example: `arewefastyet web --db-database benchmark --db-host localhost --db-password <db-password> --db-user <db-username>  
---influx-database benchmark-influx --influx-hostname localhost --influx-password <influx-password>
---influx-port <influx-port> --influx-username <influx-username> --web-api-key <web-api-key>
---web-mode production --web-port <web-port> --web-static-path ./server/static --web-template-path ./server/template
---web-webhook-config ./config.yaml`,
-		Aliases: []string{"w"},
+		Use:   "api",
+		Short: "Starts the api server of arewefastyet and the CRON service",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			err := srv.Init()
 			if err != nil {
